[REFACTOR] 안쓰는 파일 지우고 로그인, 로그아웃 함수 모듈화
